+++ CHANGELOG.md
@@ -26,6 +26,9 @@ For instructions on upgrading to newer versions, visit
 
 * \#1348 Eager loading is now supported on many-to-many relations.
 
+* \#1081 Mongoid indexes both id and type as a compound index when providing
+  `index: true` to a polymorphic belongs_to.
+
 * \#1053 Raise a `Mongoid::Errors::UnknownAttribute` instead of no method
   when attempting to set a field that is not defined and allow dynamic
   fields is false. (Cyril Mougel)